Improve structure of financial assessments code
Moved business logic to service layer and created a separate class for performing database operations on financial assessment details.
