+++ maat-court-data-api/src/main/java/gov/uk/courtdata/assessment/impl/FinancialAssessmentImpl.java
@@ -2,23 +2,14 @@ package gov.uk.courtdata.assessment.impl;
 
 import gov.uk.courtdata.assessment.mapper.FinancialAssessmentMapper;
 import gov.uk.courtdata.dto.FinancialAssessmentDTO;
-import gov.uk.courtdata.entity.FinancialAssessmentDetailsEntity;
 import gov.uk.courtdata.entity.FinancialAssessmentEntity;
 import gov.uk.courtdata.enums.FinancialAssessmentType;
-import gov.uk.courtdata.exception.MAATCourtDataException;
-import gov.uk.courtdata.model.assessment.FinancialAssessmentDetails;
-import gov.uk.courtdata.repository.FinancialAssessmentDetailsRepository;
 import gov.uk.courtdata.repository.FinancialAssessmentRepository;
 import gov.uk.courtdata.repository.HardshipReviewRepository;
 import gov.uk.courtdata.repository.PassportAssessmentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -29,16 +20,12 @@ public class FinancialAssessmentImpl {
     private final PassportAssessmentRepository passportAssessmentRepository;
     private final HardshipReviewRepository hardshipReviewRepository;
     private final FinancialAssessmentRepository financialAssessmentRepository;
-    private final FinancialAssessmentDetailsRepository financialAssessmentDetailsRepository;
 
-    public FinancialAssessmentEntity getAssessment(Integer financialAssessmentId) {
+    public FinancialAssessmentEntity find(Integer financialAssessmentId) {
         return financialAssessmentRepository.getById(financialAssessmentId);
     }
 
-    @Transactional(rollbackFor = MAATCourtDataException.class)
-    public FinancialAssessmentDTO updateAssessment(FinancialAssessmentDTO financialAssessment) {
-
-        log.info("Update Financial Assessment - Transaction Processing - Start");
+    public FinancialAssessmentEntity update(FinancialAssessmentDTO financialAssessment) {
         FinancialAssessmentEntity existingAssessment = financialAssessmentRepository.getById(financialAssessment.getId());
         existingAssessment.setFassInitStatus(financialAssessment.getFassInitStatus());
         existingAssessment.setFassFullStatus(financialAssessment.getFassFullStatus());
@@ -71,104 +58,26 @@ public class FinancialAssessmentImpl {
         } else {
             existingAssessment.setAssessmentType(FinancialAssessmentType.INIT.getValue());
         }
-
-        log.info("Updating existing financial assessment record");
-        existingAssessment = financialAssessmentRepository.save(existingAssessment);
-        log.info("Deleting stale financial assessment details");
-        deleteStaleAssessmentDetails(financialAssessment);
-        log.info("Creating/updating financial assessment detail records");
-        List<FinancialAssessmentDetailsEntity> savedDetailEntities = storeAssessmentDetails(existingAssessment, financialAssessment.getAssessmentDetailsList());
-        log.info("Update Financial Assessment - Transaction Processing - End");
-        return buildFinancialAssessmentDTO(existingAssessment, savedDetailEntities);
+        return financialAssessmentRepository.save(existingAssessment);
     }
 
-    public void deleteStaleAssessmentDetails(FinancialAssessmentDTO financialAssessment) {
-
-        List<FinancialAssessmentDetails> assessmentDetailsList = financialAssessment.getAssessmentDetailsList();
-        List<FinancialAssessmentDetailsEntity> oldAssessmentDetailsList =
-                financialAssessmentDetailsRepository.findAllByFinancialAssessmentId(financialAssessment.getId());
-
-        List<Integer> staleAssessmentDetailIds = oldAssessmentDetailsList
-                .stream()
-                .filter(oldDetails -> assessmentDetailsList
-                        .stream()
-                        .noneMatch(newDetails -> oldDetails.getCriteriaDetailId().equals(newDetails.getCriteriaDetailId())))
-                .collect(Collectors.toList())
-                .stream().map(FinancialAssessmentDetailsEntity::getId)
-                .collect(Collectors.toList());
-
-        financialAssessmentDetailsRepository.deleteAllByIdInBatch(staleAssessmentDetailIds);
-    }
-
-    public void deleteAssessment(Integer financialAssessmentId) {
+    public void delete(Integer financialAssessmentId) {
         financialAssessmentRepository.deleteById(financialAssessmentId);
     }
 
-    @Transactional(rollbackFor = MAATCourtDataException.class)
-    public FinancialAssessmentDTO createAssessment(FinancialAssessmentDTO financialAssessment) {
-
+    public FinancialAssessmentEntity create(FinancialAssessmentDTO financialAssessment) {
         FinancialAssessmentEntity assessmentEntity =
                 assessmentMapper.FinancialAssessmentDtoToFinancialAssessmentEntity(financialAssessment);
-
         assessmentEntity.setAssessmentType(FinancialAssessmentType.INIT.getValue());
+        return financialAssessmentRepository.save(assessmentEntity);
 
-        log.info("Create Financial Assessment - Transaction Processing - Start");
+//
+    }
+
+    public void setOldAssessmentReplaced(FinancialAssessmentDTO financialAssessment) {
         Integer repID = financialAssessment.getRepId();
-        log.info("Updating old assessments with REP_ID = {}", repID);
         financialAssessmentRepository.updateOldAssessments(repID);
-        log.info("Updating old passport assessments");
         passportAssessmentRepository.updateOldPassportAssessments(repID);
-        log.info("Updating old hardship reviews");
         hardshipReviewRepository.updateOldHardshipReviews(repID, financialAssessment.getId());
-        log.info("Creating new financial assessment record");
-        FinancialAssessmentEntity newAssessmentEntity = financialAssessmentRepository.save(assessmentEntity);
-        log.info("Creating new financial assessment detail records");
-        List<FinancialAssessmentDetailsEntity> assessmentDetailsEntities =
-                storeAssessmentDetails(newAssessmentEntity, financialAssessment.getAssessmentDetailsList());
-        log.info("Create Financial Assessment - Transaction Processing - End");
-        return buildFinancialAssessmentDTO(newAssessmentEntity, assessmentDetailsEntities);
-    }
-
-    public FinancialAssessmentDTO buildFinancialAssessmentDTO(FinancialAssessmentEntity assessmentEntity, List<FinancialAssessmentDetailsEntity> detailEntitiesList) {
-        FinancialAssessmentDTO newDto =
-                assessmentMapper.FinancialAssessmentEntityToFinancialAssessmentDTO(assessmentEntity);
-        newDto.setAssessmentDetailsList(
-                detailEntitiesList
-                        .stream()
-                        .map(assessmentMapper::FinancialAssessmentDetailsEntityToFinancialAssessmentDetails)
-                        .collect(Collectors.toList())
-        );
-        return newDto;
-    }
-
-    public List<FinancialAssessmentDetailsEntity> storeAssessmentDetails(FinancialAssessmentEntity financialAssessment, List<FinancialAssessmentDetails> assessmentDetails) {
-        List<FinancialAssessmentDetailsEntity> detailEntitiesList = new ArrayList<>();
-
-        List<FinancialAssessmentDetailsEntity> existingAssessmentDetails =
-                financialAssessmentDetailsRepository.findAllByFinancialAssessmentId(financialAssessment.getId());
-
-        for (FinancialAssessmentDetails detail : assessmentDetails) {
-            FinancialAssessmentDetailsEntity detailsEntity =
-                    assessmentMapper.FinancialAssessmentDetailsToFinancialAssessmentDetailsEntity(detail);
-            detailsEntity.setFinancialAssessmentId(financialAssessment.getId());
-
-            boolean exists = false;
-            if (!existingAssessmentDetails.isEmpty()) {
-                for (FinancialAssessmentDetailsEntity existingDetail : existingAssessmentDetails) {
-                    if (existingDetail.getCriteriaDetailId().equals(detailsEntity.getCriteriaDetailId())) {
-                        detailsEntity.setId(existingDetail.getId());
-                        detailsEntity.setUserModified(financialAssessment.getUserModified());
-                        exists = true;
-                        break;
-
-                    }
-                }
-            }
-            if (!exists) {
-                detailsEntity.setUserCreated(financialAssessment.getUserCreated());
-            }
-            detailEntitiesList.add(detailsEntity);
-        }
-        return financialAssessmentDetailsRepository.saveAll(detailEntitiesList);
     }
 }
