+++ maat-court-data-api/src/main/java/gov/uk/courtdata/assessment/service/FinancialAssessmentService.java
@@ -1,14 +1,21 @@
 package gov.uk.courtdata.assessment.service;
 
+import gov.uk.courtdata.assessment.impl.FinancialAssessmentDetailsImpl;
 import gov.uk.courtdata.assessment.impl.FinancialAssessmentImpl;
 import gov.uk.courtdata.assessment.mapper.FinancialAssessmentMapper;
 import gov.uk.courtdata.dto.FinancialAssessmentDTO;
+import gov.uk.courtdata.entity.FinancialAssessmentDetailsEntity;
 import gov.uk.courtdata.entity.FinancialAssessmentEntity;
+import gov.uk.courtdata.exception.MAATCourtDataException;
 import gov.uk.courtdata.model.assessment.CreateFinancialAssessment;
 import gov.uk.courtdata.model.assessment.UpdateFinancialAssessment;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -16,26 +23,65 @@ import org.springframework.stereotype.Service;
 public class FinancialAssessmentService {
 
     private final FinancialAssessmentImpl financialAssessmentImpl;
-    private final FinancialAssessmentMapper financialAssessmentMapper;
+    private final FinancialAssessmentDetailsImpl financialAssessmentDetailsImpl;
+    private final FinancialAssessmentMapper assessmentMapper;
 
-    public FinancialAssessmentDTO getAssessment(Integer financialAssessmentId) {
-        FinancialAssessmentEntity assessment = financialAssessmentImpl.getAssessment(financialAssessmentId);
-        return financialAssessmentMapper.FinancialAssessmentEntityToFinancialAssessmentDTO(assessment);
+    public FinancialAssessmentDTO find(Integer financialAssessmentId) {
+        FinancialAssessmentEntity assessmentEntity = financialAssessmentImpl.find(financialAssessmentId);
+        return buildFinancialAssessmentDTO(assessmentEntity);
     }
 
-    public FinancialAssessmentDTO updateAssessment(UpdateFinancialAssessment financialAssessment) {
+    @Transactional(rollbackFor = MAATCourtDataException.class)
+    public FinancialAssessmentDTO update(UpdateFinancialAssessment financialAssessment) {
+        log.info("Update Financial Assessment - Transaction Processing - Start");
         FinancialAssessmentDTO assessmentDTO =
-                financialAssessmentMapper.UpdateFinancialAssessmentToFinancialAssessmentDTO(financialAssessment);
-        return financialAssessmentImpl.updateAssessment(assessmentDTO);
+                assessmentMapper.UpdateFinancialAssessmentToFinancialAssessmentDTO(financialAssessment);
+        log.info("Updating existing financial assessment record");
+        FinancialAssessmentEntity assessmentEntity = financialAssessmentImpl.update(assessmentDTO);
+        log.info("Deleting stale financial assessment details");
+        financialAssessmentDetailsImpl.deleteStaleAssessmentDetails(assessmentDTO);
+        log.info("Updating financial assessment detail records");
+        List<FinancialAssessmentDetailsEntity> assessmentDetailsEntities =
+                financialAssessmentDetailsImpl.save(assessmentEntity, assessmentDTO.getAssessmentDetailsList());
+        log.info("Update Financial Assessment - Transaction Processing - End");
+        return buildFinancialAssessmentDTO(assessmentEntity, assessmentDetailsEntities);
     }
 
-    public void deleteAssessment(Integer financialAssessmentId) {
-        financialAssessmentImpl.deleteAssessment(financialAssessmentId);
+    public void delete(Integer financialAssessmentId) {
+        financialAssessmentImpl.delete(financialAssessmentId);
     }
 
-    public FinancialAssessmentDTO createAssessment(CreateFinancialAssessment financialAssessment) {
+    @Transactional(rollbackFor = MAATCourtDataException.class)
+    public FinancialAssessmentDTO create(CreateFinancialAssessment financialAssessment) {
+        log.info("Create Financial Assessment - Transaction Processing - Start");
         FinancialAssessmentDTO assessmentDTO =
-                financialAssessmentMapper.CreateFinancialAssessmentToFinancialAssessmentDTO(financialAssessment);
-        return financialAssessmentImpl.createAssessment(assessmentDTO);
+                assessmentMapper.CreateFinancialAssessmentToFinancialAssessmentDTO(financialAssessment);
+        log.info("Creating new financial assessment record");
+        FinancialAssessmentEntity assessmentEntity = financialAssessmentImpl.create(assessmentDTO);
+        log.info("Creating new financial assessment detail records");
+        List<FinancialAssessmentDetailsEntity> assessmentDetailsEntities =
+                financialAssessmentDetailsImpl.save(assessmentEntity, assessmentDTO.getAssessmentDetailsList());
+        log.info("Setting outdated records as replaced");
+        financialAssessmentImpl.setOldAssessmentReplaced(assessmentDTO);
+        log.info("Create Financial Assessment - Transaction Processing - End");
+        return buildFinancialAssessmentDTO(assessmentEntity, assessmentDetailsEntities);
+    }
+
+    public FinancialAssessmentDTO buildFinancialAssessmentDTO(FinancialAssessmentEntity assessmentEntity, List<FinancialAssessmentDetailsEntity> detailEntitiesList) {
+        FinancialAssessmentDTO newDto =
+                assessmentMapper.FinancialAssessmentEntityToFinancialAssessmentDTO(assessmentEntity);
+        if (detailEntitiesList != null) {
+            newDto.setAssessmentDetailsList(
+                    detailEntitiesList
+                            .stream()
+                            .map(assessmentMapper::FinancialAssessmentDetailsEntityToFinancialAssessmentDetails)
+                            .collect(Collectors.toList())
+            );
+        }
+        return newDto;
+    }
+
+    public FinancialAssessmentDTO buildFinancialAssessmentDTO(FinancialAssessmentEntity assessmentEntity) {
+        return buildFinancialAssessmentDTO(assessmentEntity, null);
     }
 }
